+++ docs/advanced/whatsnew.rst
@@ -6,12 +6,13 @@ What's New
 
 .. note:: If you already have an old version of ``flocker-cli`` installed, delete the install script and directory before installing the newer version.
 
-Next Release
-============
+v0.4
+====
 
 * New :ref:`REST API<api>` for managing datasets.
 * Applications can now be configured with a :ref:`restart policy<restart configuration>`.
 * Volumes can now be configured with a :ref:`maximum size<volume configuration>`.
+* Documentation now includes :ref:`instructions for installing flocker-node on CentOS 7<centos-7-install>`.
 
 v0.3.2
 ======